Unify anomalib to LightningEngine
