+++ src/otx/v2/adapters/__init__.py
@@ -9,7 +9,6 @@ ADAPTERS = [
     "openvino",
     "torch",
     "torch.lightning",
-    "torch.lightning.anomalib",
     "torch.mmengine",
     "torch.mmengine.mmdeploy",
     "torch.mmengine.mmpretrain",
