--- src/otx/v2/adapters/torch/lightning/anomalib/modules/__init__.py
@@ -1,4 +1,0 @@
-"""OTX Adapters - Anomalib."""
-
-# Copyright (C) 2021-2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
