+++ src/otx/v2/adapters/torch/lightning/modules/datasets/__init__.py
@@ -3,6 +3,10 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+# [TODO]: This should change with anomalib v1.0.
+from otx.v2.adapters.torch.lightning.registry import DATASETS
+
+from .anomaly_dataset import OTXAnomalyDataModule, OTXAnomalyDataset
 from .pipelines import MultipleInputsCompose, Pad, ResizeLongestSide
 from .visual_prompting_dataset import OTXVisualPromptingDataModule, OTXVisualPromptingDataset, get_transform
 
@@ -13,4 +17,10 @@ __all__ = [
     "MultipleInputsCompose",
     "Pad",
     "ResizeLongestSide",
+    "OTXAnomalyDataModule",
+    "OTXAnomalyDataset",
+    "DATASETS",
 ]
+
+DATASETS.register_module(name="visual_prompting", module=OTXVisualPromptingDataset)
+DATASETS.register_module(name="anomaly", module=OTXAnomalyDataset)
