+++ src/otx/v2/adapters/torch/lightning/modules/datasets/utils/__init__.py
@@ -1,2 +1,4 @@
+"""OTX Ligthning adapter modules - utils functions."""
+
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
