+++ src/otx/v2/adapters/torch/lightning/modules/models/__init__.py
@@ -3,10 +3,9 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from otx.v2.adapters.torch.lightning.registry import LightningRegistry
+from otx.v2.adapters.torch.lightning.registry import MODELS
 
+from . import anomaly
 from .visual_prompters import SegmentAnything
 
-# NOTE: Register the model with the Registry to make it available via the config API.
-MODELS = LightningRegistry()
-MODELS.register_module(name="SAM", module=SegmentAnything)
+__all__ = ["anomaly", "SegmentAnything", "MODELS"]
