--- tests/v2/unit/adapters/torch/lightning/anomalib/modules/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
