--- tests/v2/unit/adapters/torch/lightning/anomalib/modules/data/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (C) 2021-2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
